arch: move the timestamp and yield backends into C files
jent_get_nstime() and jent_yield() were inline in the headers, which
pulled every platform header - <windows.h>, <x86intrin.h>, the Mach
headers, <linux/timex.h> - into the entropy-collection core through
jitterentropy-internal.h. That does not survive the -O0 the kernel build
compiles the core at. The backends now live in arch-timer.c and
arch-sched.c behind one out-of-line call, which delays the measurement
uniformly rather than removing jitter from it, and kernel and user space
read the same counter wherever the instruction set offers one.

The generic user-space fallback takes CLOCK_MONOTONIC: CLOCK_REALTIME is
stepped and slewed by adjtime/NTP, which is external interference rather
than jitter the CPU produced, and it can run backwards.
