+++ arch/jitterentropy-arch-sched.c
@@ -71,11 +71,23 @@
 #  include <x86intrin.h>
 # endif
 # define JENT_ARCH_SCHED_PAUSE_X86
+#elif defined(_M_ARM64) || defined(_M_ARM)
+/*
+ * Windows on ARM. MSVC defines neither __aarch64__ nor __arm__, so without
+ * these two macros the branch below did not match and a Windows-on-ARM build
+ * emitted no yield hint at all - the same pair jitterentropy-arch-timer.c
+ * matches for its QueryPerformanceCounter back-end. The hint goes through the
+ * intrinsic because MSVC rejects inline asm on ARM targets entirely.
+ */
+# include <intrin.h>
+# define JENT_ARCH_SCHED_PAUSE_ARM_INTRIN
 #elif defined(__aarch64__) || \
       (defined(__arm__) && defined(__ARM_ARCH) && __ARM_ARCH >= 7)
 # define JENT_ARCH_SCHED_PAUSE_ARM
 #elif defined(__powerpc) || defined(__powerpc__)
 # define JENT_ARCH_SCHED_PAUSE_POWERPC
+#elif defined(__riscv)
+# define JENT_ARCH_SCHED_PAUSE_RISCV
 #endif
 
 #endif /* LINUX_KERNEL */
@@ -84,10 +96,22 @@ void jent_yield(void)
 {
 #if defined(JENT_ARCH_SCHED_PAUSE_X86)
 	_mm_pause();
+#elif defined(JENT_ARCH_SCHED_PAUSE_ARM_INTRIN)
+	__yield();
 #elif defined(JENT_ARCH_SCHED_PAUSE_ARM)
 	__asm__ __volatile__("yield" ::: "memory");
 #elif defined(JENT_ARCH_SCHED_PAUSE_POWERPC)
 	__asm__ __volatile__("or 27,27,27" ::: "memory");
+#elif defined(JENT_ARCH_SCHED_PAUSE_RISCV)
+	/*
+	 * The Zihintpause "pause" hint, spelled as its raw encoding rather than
+	 * as the mnemonic. The assembler only accepts "pause" when the
+	 * extension is enabled in the target's -march string, which it usually
+	 * is not; the instruction itself is architecturally a FENCE hint and so
+	 * executes as a no-op on hardware that predates Zihintpause. Emitting
+	 * the encoding directly therefore works everywhere.
+	 */
+	__asm__ __volatile__(".insn i 0x0F, 0, x0, x0, 0x010" ::: "memory");
 #endif
 
 #if defined(JENT_ARCH_SCHED_OS_WINDOWS)
